refactor: extract workflow task detail read model
